requirements.txt
streamlit>=1.31.0
tensorflow>=2.13.0
numpy>=1.24.3
pandas>=2.0.3
langchain>=0.1.0
langchain-community>=0.0.13
python-dotenv>=1.0.0
google-generativeai>=0.3.1
requests>=2.31.0
psutil>=5.9.8
unstructured>=0.11.0
python-magic-bin>=0.4.14  # for file type detection
pdf2image>=1.16.3  # for PDF processing
pytesseract>=0.3.10  # for OCR
pillow>=10.2.0  # for image processing
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
watchdog
pathlib>=1.0.1
mimetypes>=0.1.0
